+++ src/main/java/org/sopt/kareer/domain/roadmap/facade/ActionItemFacade.java
@@ -4,8 +4,8 @@ import lombok.RequiredArgsConstructor;
 import org.sopt.kareer.domain.roadmap.dto.response.ActionItemListResponse;
 import org.sopt.kareer.domain.roadmap.dto.response.ActionItemResponse;
 import org.sopt.kareer.domain.roadmap.entity.enums.ActionItemType;
-import org.sopt.kareer.domain.roadmap.service.ActionItemCommandService;
-import org.sopt.kareer.domain.roadmap.service.ActionItemQueryService;
+import org.sopt.kareer.domain.roadmap.service.actionitem.ActionItemCommandService;
+import org.sopt.kareer.domain.roadmap.service.actionitem.ActionItemQueryService;
 import org.sopt.kareer.domain.roadmap.service.dto.response.ActionItemDetail;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
