+++ src/main/java/org/sopt/kareer/domain/roadmap/repository/PhaseActionGuidelineRepository.java
@@ -1,6 +1,6 @@
 package org.sopt.kareer.domain.roadmap.repository;
 
-import org.sopt.kareer.domain.roadmap.entity.PhaseActionGuideline;
+import org.sopt.kareer.domain.roadmap.entity.phaseaction.PhaseActionGuideline;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
